+++ Source/Game/Spells/SpellInfo.cs
@@ -4911,7 +4911,7 @@ namespace Game.Spells
         public SpellEffectName Effect;
         public AuraType ApplyAuraName;
         public uint ApplyAuraPeriod;
-        public int BasePoints;
+        public float BasePoints;
         public float RealPointsPerLevel;
         public float PointsPerResource;
         public float Amplitude;
